工作流库0216/T37-同花顺excel导入财务指标/requirements.txt
# T37-同花顺Excel导入财务指标 依赖列表
# Python 3.8+

# 数据处理
numpy>=1.21.0
pandas>=1.3.0
openpyxl>=3.0.9

# 数据库连接
sqlalchemy>=1.4.0
pymysql>=1.0.0
psycopg2-binary>=2.9.0

# 同花顺iFinD API（需要单独安装，非PyPI）
# iFinDPy

# Excel自动化（Windows专用）
pywin32>=300

# 日期处理
python-dateutil>=2.8.0

# 日志和工具
tqdm>=4.62.0

# 可选：数据验证
pydantic>=1.8.0
